+++ portfolio/src/main/java/com/google/sps/servlets/Testimonial.java
@@ -1,0 +1,14 @@
+package com.google.sps.servlets;
+
+public class Testimonial {
+    private String text;
+    private String name;
+    private String title;
+
+    public Testimonial(String text, String name, String title){
+        this.text = text;
+        this.name = name;
+        this.title = title;
+    }
+
+}
