+++ portfolio/src/main/java/com/google/sps/servlets/Testimonials.java
@@ -24,16 +24,3 @@ public class Testimonials {
       return "Testimonials [ arrTestimonials: "+arrTestimonials+",]"; 
    }  
 }
-
-class Testimonial {
-    private String text;
-    private String name;
-    private String title;
-
-    public Testimonial(String text, String name, String title){
-        this.text = text;
-        this.name = name;
-        this.title = title;
-    }
-
-}
